Split some more steps in their own file
